separate out sse logic
